requirements.txt
blinker==1.6.2
boto3==1.28.26
botocore==1.31.26
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.6
Flask==2.3.2
Flask-Cors==4.0.0
Flask-SQLAlchemy==3.0.5
greenlet==2.0.2
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
jmespath==1.0.1
MarkupSafe==2.1.3
psycopg2-binary==2.9.7
python-dateutil==2.8.2
python-dotenv==1.0.0
requests==2.31.0
s3transfer==0.6.1
six==1.16.0
SQLAlchemy==2.0.19
typing_extensions==4.7.1
urllib3==1.26.16
Werkzeug==2.3.6
yfinance
